==> 02_scripts/01_base_model/03_base_model.Rmd <==
---
title: "Base Model"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Model Compas Dataset

## Goal

After executing this script, the base model (logistic regression) for the COMPAS data set should be completed.

## Install packages

```{r}
pacman::p_load(tidyverse, 
               here,
               tidymodels,
               knitr,
               kableExtra,
               magick)

tidymodels_prefer()
```

## Load data

```{r}
data = readRDS(here("01_data", "02_data_processed", "01_base_model",
                    "03_data.rds"))
```

## Data preparation
# Final Data Transformation
```{r}
recipe = recipe(compas ~ .,
                data = data) %>% 
  step_dummy(all_nominal_predictors())
```

## Modelling (Logistic Regression)
```{r}
model = logistic_reg() %>% 
  set_engine("glm") %>% 
  set_mode("classification")

workflow = workflow() %>% 
  add_model(model) %>% 
  add_recipe(recipe)

model_fit = 
  workflow %>% 
  fit(data = data)

estimation = tidy(model_fit, exponentiate = TRUE)

estimation %>% 
  slice(-1) %>% 
  mutate(estimate_percent = percent((estimate - 1), digits = 1)) %>% 
  relocate(estimate_percent, .after = estimate)

estimation_nice = estimation %>% 
    slice(-1) %>% 
    mutate(term = case_when(
      term == "age" ~ "Alter",
      term == "days_in_jail" ~ "Tage im Gefängnis",
      term == "juv_fel_count" ~ "Straftaten (jugendl.)",
      term == "juv_misd_count" ~ "Ordnungswidr. (jugendl.)",
      term == "juv_other_count" ~ "Andere Anzeigen (jugendl.)",
      term == "priors_count" ~ "Anzahl Straftaten",
      term == "sex_Female" ~ "Geschlecht: Weiblich",
      term == "race_African.American" ~ "Race: Schwarz",
      term == "race_Hispanic" ~ "Race: Hispanisch",
      term == "race_Other" ~ "Race: Andere",
      term == "charge_degree_M" ~ "Schwere der Straftat: Ordnungswidr.",
      term == "two_year_recid_yes" ~ "Rückfälligkeit: Ja",
      .default = term
        )
      ) %>% 
    column_to_rownames(var = "term") %>% 
    mutate(estimate_percent = percent(estimate-1, accuracy = 0.1)) %>% 
    relocate(estimate_percent, .after = estimate) %>% 
    rename("Schätzer" = estimate,
           "Schätzer (in Prozent)" = estimate_percent,
           Standardfehler = std.error,
           "z-Wert" = statistic,
           "p-Wert" = p.value)

kable(estimation_nice, booktabs = TRUE, 
      align = "c", 
      digits = c(3, 1, 3, 2, 4)) %>% 
  row_spec(seq(1, nrow(estimation_nice), 2), background = "#f8f8f8") %>% 
  kable_styling() %>% 
  save_kable(file = here("03_outputs", "01_base_model", "09_table_estimates.png"),
             density = 1000)
```


```{r}
saveRDS(model_fit, 
        file = here("01_data", "02_data_processed", "01_base_model",
                    "04_model_fit.rds")
        )
```
